Move vaccin kpi section to own component
